Replace BaseIssue with Issue interface
