+++ src/app/core/models/issue.model.ts
@@ -2,49 +2,111 @@ import { Team } from './team.model';
 import { TesterResponse } from './tester-response.model';
 import { IssueComment } from './comment.model';
 import { IssueDispute } from './issue-dispute.model';
+import { GithubIssue, GithubLabel } from './github-issue.model';
+import { GithubComment } from './github-comment.model';
+import { TeamResponseTemplate } from './templates/team-response-template.model';
+import { TesterResponseTemplate } from './templates/tester-response-template.model';
+import { TutorModerationIssueTemplate } from './templates/tutor-moderation-issue-template.model';
+import { TutorModerationTodoTemplate } from './templates/tutor-moderation-todo-template.model';
+import * as moment from 'moment';
 
-export interface Issue {
+export class Issue {
+
+  /** Basic Fields */
   readonly id: number;
   readonly created_at: string;
   title: string;
-  type: string;
+  description: string;
+
+  /** Fields derived from Labels */
   severity: string;
-  description?: string;
+  type: string;
   responseTag?: string;
-  assignees?: string[];
   duplicated?: boolean;
-  duplicateOf?: number;
   status?: string;
+  pending?: string;
+  unsure?: boolean;
   teamAssigned?: Team;
+
+  /** Depending on the phase, assignees attribute can be derived from Github's assignee feature OR from the Github's issue description */
+  assignees?: string[];
+
+  /** Fields derived from parsing of Github's issue description */
+  duplicateOf?: number;
   todoList?: string[];
   teamResponse?: string;
   tutorResponse?: string;
   testerResponses?: TesterResponse[];
   issueComment?: IssueComment; // Issue comment is used for Tutor Response and Tester Response
   issueDisputes?: IssueDispute[];
-  pending?: string;
-  unsure?: boolean;
+
+  protected constructor(githubIssue: GithubIssue) {
+    /** Basic Fields */
+    this.id = +githubIssue.number;
+    this.created_at = moment(githubIssue.created_at).format('lll');
+    this.title = githubIssue.title;
+    this.description = githubIssue.body;
+
+    /** Fields derived from Labels */
+    this.severity = githubIssue.findLabel(GithubLabel.LABELS.severity);
+    this.type = githubIssue.findLabel(GithubLabel.LABELS.type);
+    this.responseTag = githubIssue.findLabel(GithubLabel.LABELS.response);
+    this.duplicated = !!githubIssue.findLabel(GithubLabel.LABELS.duplicated, false);
+    this.status = githubIssue.findLabel(GithubLabel.LABELS.status);
+    this.pending = githubIssue.findLabel(GithubLabel.LABELS.pending);
+  }
+
+  public static createPhaseBugReportingIssue(githubIssue: GithubIssue): Issue {
+    return new Issue(githubIssue);
+  }
+
+  public static createPhaseTeamResponseIssue(githubIssue: GithubIssue, githubComments: GithubComment[],
+                                             teamData: Team): Issue {
+    const issue = new Issue(githubIssue);
+    const template = new TeamResponseTemplate(githubComments);
+
+    issue.teamAssigned = teamData;
+    issue.issueComment = template.comment;
+    issue.teamResponse = template.teamResponse !== undefined ? template.teamResponse.content : undefined;
+    issue.duplicateOf = template.duplicateOf !== undefined ? template.duplicateOf.issueNumber : undefined;
+    issue.duplicated = issue.duplicateOf !== undefined && issue.duplicateOf !== null;
+    issue.assignees = githubIssue.assignees.map(assignee => assignee.login);
+
+    return issue;
+  }
+
+  public static createPhaseTesterResponseIssue(githubIssue: GithubIssue, githubComments: GithubComment[]): Issue {
+    const issue = new Issue(githubIssue);
+    const template = new TesterResponseTemplate(githubComments);
+    issue.issueComment = template.comment;
+    issue.teamResponse = template.teamResponse !== undefined ? template.teamResponse.content : undefined;
+    issue.testerResponses = template.testerResponse.testerResponses;
+    return issue;
+  }
+
+  public static createPhaseModerationIssue(githubIssue: GithubIssue, githubComments: GithubComment[],
+                                           teamData: Team): Issue {
+    const issue = new Issue(githubIssue);
+    const issueTemplate = new TutorModerationIssueTemplate(githubIssue);
+    const todoTemplate = new TutorModerationTodoTemplate(githubComments);
+    const disputes = todoTemplate.moderation.disputesToResolve.map((dispute, i) => {
+      dispute.description = issueTemplate.dispute.disputes[i].description;
+      return dispute;
+    });
+
+    issue.teamAssigned = teamData;
+    issue.issueComment = todoTemplate.comment;
+    issue.description = issueTemplate.description.content;
+    issue.teamResponse = issueTemplate.teamResponse.content;
+    issue.issueDisputes = disputes;
+    issue.todoList = todoTemplate.moderation.todoList;
+    return issue;
+  }
 }
 
 export interface Issues {
   [id: number]: Issue;
 }
-
-/**
- * The types of labels in Issue must follow the format of `Type`.`Value`.
- * Where `Type` represent the type of the label. (e.g. severity, type, response)
- * And `Value` represent the value that is associated to the `Type` (e.g. for severity Type, it could be Low, Medium, High)
- */
-export const LABELS = ['severity', 'type', 'response', 'duplicate', 'status', 'pending', 'unsure'];
-
-export const labelsToAttributeMapping = {
-  'severity': 'severity',
-  'type': 'type',
-  'response': 'responseTag',
-  'status': 'status',
-  'pending': 'pending',
-  'unsure': 'unsure'
-};
 
 export const SEVERITY_ORDER = { '-': 0 , Low: 1, Medium: 2, High: 3 };
 
@@ -75,12 +137,6 @@ export enum RESPONSE {
   CannotReproduce = 'CannotReproduce',
 }
 
-export const ISSUE_LABELS = {
-  severity: Object.keys(SEVERITY),
-  type: Object.keys(TYPE),
-  response: Object.keys(RESPONSE),
-};
-
 export const IssuesFilter = {
   phaseBugReporting: {
     Student: 'FILTER_BY_CREATOR',
@@ -108,15 +164,3 @@ export enum RespondType {
   teamResponse = 'teamResponse',
   tutorResponse = 'tutorResponse',
 }
-
-export const phaseTeamResponseDescriptionTemplate = new RegExp('(?<header># Description|# Team\'s Response|## State the duplicated issue ' +
-  'here, if any)\\s+(?<description>[\\s\\S]*?)(?=# Team\'s Response|## State the duplicated issue here, if any|$)', 'gi');
-
-export const phaseTesterResponseDescriptionTemplate = new RegExp('(?<header># Description|# Team\'s Response|## State ' +
-  'the duplicated issue here, if any|# Items for the Tester to Verify)\\s+(?<description>[\\s\\S]*?)(?=# Team\'s Response' +
-  '|## State the duplicated issue here, if any|# Items for the Tester to Verify|$)', 'gi');
-
-export const phaseModerationDescriptionTemplate = new RegExp('(?<header># Description|# Team\'s Response|## State the duplicated issue ' +
-  'here, if any|# Disputes)\\s+' +
-  '(?<description>[\\s\\S]*?)(?=# Team\'s Response|## State the duplicated issue here, if any|' +
-  '# Disputes|$)', 'gi');
